requirements.txt
requests==2.27.0
jupyterlab==3.2.5
biopython==1.79
pandas==1.3.5
matplotlib==3.5.1
scipy==1.7.3
plotly==5.7.0
scikit-learn==1.1.0
kaleido==0.2.1
seaborn==0.11.2